% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrupar_por.R
\name{agrupar_por}
\alias{agrupar_por}
\title{Agrupa datos de objeto xts según intervalo de tiempo dado}
\usage{
agrupar_por(xts, int_salida = "MENSUAL")
}
\arguments{
\item{xts}{xts de entrada. La columna de datos a agrupar debe llamarse "energia".}

\item{int_salida}{Cadena de caracteres para especificar el nuevo intervalo de tiempo entre datos.
Puede ser: "HORA", "DIARIO", "MENSUAL", según lo detallado en la descripción de la función}
}
\value{
xts con los datos agrupados según el nuevo intervalo de tiempo, con una sola columna llamada "energia".
}
\description{
Agrupa datos de objeto xts según intervalo de tiempo dado. El intervalo de tiempo de los datos del
xts de salida pueden ser cada una hora "HORA", cada un día "DIARIO" o cada un mes "MENSUAL". El intervalo
de datos del xts de entrada debe ser menor que el de salida. Por ejemplo, se puede pasar de datos cada
15 minutos a datos cada una hora, o de datos cada un día a datos cada un mes.
}
\examples{
# #Importa datos de medidor cada 15 minutos
# datos_15min <- importar_optimum(usuario="usuario_valido", password="password_valido",
#                                 medidores=",CIR0141449301", fecha_inicio = "2019-03-01",
#                                 fecha_fin = "2020-03-15", int_entrada = "15MIN", consumo = TRUE)
# #Agrupa los datos cada 1 mes
# datos_mes <- agrupar_por(datos_15min)
}
\author{
Daniel G. Paniagua
}
