% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descarga_optimum.R
\name{descarga_optimum}
\alias{descarga_optimum}
\title{Descarga datos de medidores de energía de Optimum}
\usage{
descarga_optimum(
  usuario,
  password,
  medidores = "GEN1HORA",
  fecha_inicio = "2019-01-01",
  fecha_fin = "2019-01-02",
  channel = "Active Import Inc - 0"
)
}
\arguments{
\item{usuario}{Cadena de caracteres. Nombre de usuario para login en Optimum.}

\item{password}{Cadena de caracteres. Contraseña para login en Optimum.}

\item{medidores}{Cadena de caracteres.Se puede ingresar el nombre de alguno de los grupos predefinidos
que se detallan a continuación. En caso de no querer descargar ninguno de estos grupos predefinidos, se puede
detallar una cadena con el/los número/s de medidor/es con una coma delante.
\cr Por ejemplo: ',CIR0141437453,CIR0141449390,CIR0141449501'. \cr Los grupos predefinidos son: \itemize{
 \item{medidores="TODOS":}{ Todos los medidores Circutor de Optimum. Son 263 medidores. Todos se descargan del
 canal "Active Import Inc - 0", excepto la planta de piso y un medidor de generación cada una hora (ver detalle en siguientes ítems).}
 \item{medidores="GEN15MIN":}{ Corresponde a 6 medidores de generación con datos cada 15 minutos. Todos en
 canal "Active Import Inc - 0". }
 \item{medidores="GEN1HORA":}{ Corresponde a 5 medidores de generación con datos cada 1 hora con datos en
 canal "Active Import Inc - 0" y a uno (CIR0141668286) con datos en canal 'Active Export Inc - 0'. }
 \item{medidores="DEM15MIN":}{ Corresponde a 230 medidores de consumo con datos cada 15 minutos. Todos los datos
  en se descargan de canal "Active Import Inc - 0" (hasta actual estado de depuración). }
 \item{medidores="DEM1HORA":}{ Corresponde a 20 medidores de consumo con datos cada 1 hora. Todos los datos
  en se descargan de canal "Active Import Inc - 0" (hasta actual estado de depuración). }
 \item{medidores="PLANTAPISO":}{ Corresponde a datos de la planta de piso, los datos están
 en canal "Sum Li Active Power + (QI+QIV) (time integral 5) - 0".}}}

\item{fecha_inicio}{Cadena de caracteres. Fecha de inicio de los datos a descargar, en formato "YYYY-MM-DD".}

\item{fecha_fin}{Cadena de caracteres. Fecha final de los datos a descargar, en formato "YYYY-MM-DD".}

\item{channel}{Cadena de caracteres. Canal del medidor que contiene los datos a descargar. Los más utilizados son:
"Active Import Inc - 0" (la mayoría está en este canal), 'Active Export Inc - 0', "Sum Li Active Power + (QI+QIV) (time integral 5) - 0"
(solo planta de piso). Ver detalle en descripción de parámetro "medidores".}
}
\value{
data.frame con los datos descargados, con la misma disposición que la tabla descargada desde la web de Optimum.
}
\description{
Descarga datos de medidores de energía de Optimum (Armstrong) y los devuelve como data.frame.
}
\examples{
# #Descarga datos de 230 medidores con datos cada 15 minutos, por el período de 13 meses.
# #En la prueba realizada se descargaron 7.841.964 datos en 973 segundos (16min 12seg aprox).
# #Se deben ingresar usuario y password validos (no incluidos en el ejemplo).
#
# datos_dem_190 <- descarga_optimum(usuario="usuario valido", password="password valido",
#                                   medidores="DEM15MIN", fecha_inicio = "2019-03-01",
#                                   fecha_fin = "2020-04-30")
#
# #Descarga datos de 1 medidor cada una hora por un mes.
# datos_1 <- descarga_optimum(usuario="usuario valido", password="password valido",
#                             medidores=",CIR0141449301", fecha_inicio = "2019-03-01",
#                             fecha_fin = "2019-03-31")
}
\author{
Daniel G. Paniagua
}
