% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importar_epw.R
\name{importar_epw}
\alias{importar_epw}
\title{Importa datos de radiación y temperatura de archivo meteorológico epw}
\usage{
importar_epw(rutaepw)
}
\arguments{
\item{rutaepw}{Cadena de caracteres con ruta del archivo epw.}
}
\value{
Un objeto xts con dos columnas: \itemize{
 \item{G0: }{Radiación solar global en el plano horizontal (kWh/m2)}
 \item{Ta: }{Temperatura media (ºC)}
}
}
\description{
Importa datos de radiación solar global en el plano horizontal y
de temperatura de un archivo epw.
}
\details{
La función debe asignar un año a los datos, ya que en un archivo epw
los datos de los distintos meses generalmente corresponden a distintos años no
consecutivos. El año es asignado por el método "data()" de la clase
Epw (ver \link[eplusr]{Epw}), que tiene en cuenta no generar errores por años bisiestos.
Si bien el método permite asignar el año, se prefiere no hacerlo para no generar
este tipo de errores. Puede suceder que quede un dato correspondiente al año siguiente,
Esto es porque la hora del dato siempre corresponde a la correspondiente al fin del
intervalo de tiempo, generando un dato del año siguiente a las cero horas.
\cr. Por otra parte, la hora en el índice del objeto xts devuelto incluye el huso horario
contenido en el archivo epw.
}
\examples{
# #Importa radiacion global horizontal y temperatua del archivo epw. El archivo especificado
# #debe estar en el directorio de trabajo o bien dar la ruta completa.
#
# rad_dia <- importar_epw("ARG_ROSARIO-AERO_874800_IW2.EPW")
}
\references{
\itemize{
 \item{}{Jia H (2020). eplusr: A Toolkit for Using EnergyPlus in R.
 R package version 0.12.0, https://CRAN.R-project.org/package=eplusr.}
}
}
\author{
Daniel G. Paniagua
}
