% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importar_optimum.R
\name{importar_optimum}
\alias{importar_optimum}
\title{Importa y descarga datos de generación de energía de Optimum}
\usage{
importar_optimum(
  file = NULL,
  usuario,
  password,
  medidores = "GEN1HORA",
  fecha_inicio = "2019-01-01",
  fecha_fin = "2019-01-02",
  channel = "Active Import Inc - 0",
  int_entrada = "HORA",
  int_salida = NULL,
  na = "INTERPOLAR",
  mult_us = FALSE,
  consumo = FALSE,
  indicar_na = FALSE
)
}
\arguments{
\item{file}{Cadena de caracteres, data.frame o NULL. Ruta del archivo a importar o nombre del data.frame.
Puede ser xls, xlsx o csv. Para csv, el caracter separador debe ser coma. Para descargar datos directamente
de Optimum debe ser NULL (por defecto).}

\item{usuario}{Cadena de caracteres. Nombre de usuario para login en Optimum. Solo se utiliza si los
datos se descargan directamente de Optimum. Parámetro de \link[OESolar]{descarga_optimum}.}

\item{password}{Cadena de caracteres. Contraseña para login en Optimum. Solo se utiliza si los
datos se descargan directamente de Optimum. Parámetro de \link[OESolar]{descarga_optimum}.}

\item{medidores}{Ver detalles en \link[OESolar]{descarga_optimum}.}

\item{fecha_inicio}{Cadena de caracteres. Fecha de inicio de los datos a descargar, en formato "YYYY-MM-DD". Solo se utiliza si los
datos se descargan directamente de Optimum. Parámetro de \link[OESolar]{descarga_optimum}.}

\item{fecha_fin}{Cadena de caracteres. Fecha final de los datos a descargar, en formato "YYYY-MM-DD". Solo se utiliza si los
datos se descargan directamente de Optimum. Parámetro de \link[OESolar]{descarga_optimum}.}

\item{channel}{Cadena de caracteres. Canal del medidor que contiene los datos a descargar. Solo se utiliza si los
datos se descargan directamente de Optimum. Parámetro de \link[OESolar]{descarga_optimum}.}

\item{int_entrada}{Cadena de caracteres. "15MIN" u "HORA". Corresponde al
intervalo de tiempo entre los datos de entrada. No utilizar datos que mezclen estos valores.}

\item{int_salida}{Cadena de caracteres. "15MIN", "HORA", "DIARIO" o "MENSUAL". Corresponde al
intervalo de tiempo entre los datos que se desee para el objeto xts de salida.}

\item{na}{Cadena de caracteres. NULL, "CERO" o "INTERPOLAR", según cómo se quieran rellenar
los huecos: \itemize{
 \item{NULL: }{no se rellenan los huecos.}
 \item{"CERO": }{rellena todos los huecos con ceros.}
 \item{"INTERPOLAR": }{ rellena huecos interpolando los datos faltantes con los datos anterior y posterior al hueco.
 Si \code{consumo=TRUE}, interpola para todos los horarios. Si \code{consumo=FALSE}, interpola en horarios diurnos
 y completa con cero en horarios nocturnos.}
 }}

\item{mult_us}{Lógico. Si es TRUE, los datos a importar corresponden a más de un usuario.}

\item{consumo}{Lógico. TRUE si los datos son de consumo. Si es FALSE (cuando los datos son de generación),
la interpolación para NA en horario nocturno no se hace y se completa con ceros.}

\item{indicar_na}{Lógico. Si es TRUE, los datos agregados se identifican con un "1" en una columna adicional
llamada "rellenado".}
}
\value{
\itemize{
 \item{Si mult_us=FALSE: }{Un objeto xts con la fecha como índice y el valor de energía en el campo "energia"}
 \item{Si mult_us=TRUE: }{Un objeto "list" cuyos elementos son "xts" como los descriptos en ítem anterior
 y cada uno corresponde a un usuario distinto}
 }
}
\description{
Importa datos de generación de energía del sistema Optimum, descargándolos directamente, de archivos
en formato csv o xls/xlsx descargados desde la web, o de data.frame devuelto por descarga_optimum().
}
\examples{
#
# ################A partir de archivo xlsx##############
# gen_data_mes <- importar_optimum("D:/Pruebas R/prodGCPV/PANEL SALA VELATORIA 1 2019.xlsx")
#
# gen_data_dia <- importar_optimum("D:/Pruebas R/prodGCPV/PANEL SALA VELATORIA 1 2019.xlsx",
#                                  int_salida = "DIARIO")
#
# gen_data_hora <- importar_optimum("D:/Pruebas R/prodGCPV/PANEL SALA VELATORIA 1 2019.xlsx",
#                                   int_salida = "HORA")
#
#
# ################A partir de data.frame devuelto por descarga_optimum##############
# #Descarga datos de 230 medidores con datos cada 15 minutos, por el periodo de 13 meses. En la
# #prueba realizada se descargaron 7.841.964 datos en 973 segundos (16min 12seg aprox). Se deben
# #ingresar usuario y password validos (no incluidos en el ejemplo)
# prueba_dem_year <- descarga_optimum(usuario="usuario valido", password = "password valido",
#                                    medidores = "DEM15MIN", fecha_inicio = "2019-03-01",
#                                    fecha_fin = "2020-04-30")
#
# #Importa los datos descargados, pasandolos a un "list" donde cada elemento corresponde a un
# #medidor distinto y es un objeto "xts". Para este ejemplo, la funcion se ejecuto
# #en 385 segundos (6min 25seg aprox)
# list.prueba_dem_year <- importar_optimum(file=prueba_dem_year, int_entrada = "15MIN",
#                                         mult_us = TRUE, consumo = TRUE)
#
#
# ################Descargando datos directamente##############
# #El ejemplo anterior se puede ejecutar usando directamente la funcion importar_optimum.
# #Para este ejemplo, la funcion se ejecuto en 1433 segundos (23min 53seg aprox).:
# list.prueba_dem_year <- importar_optimum(usuario="usuario_valido", password = "password_valido",
#                                         medidores = "DEM15MIN", fecha_inicio = "2019-03-01",
#                                         fecha_fin = "2020-04-30",int_entrada = "15MIN",
#                                         mult_us = TRUE, consumo = TRUE)
}
\author{
Daniel G. Paniagua
}
