% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_huecos.R
\name{info_huecos}
\alias{info_huecos}
\title{Devuelve data.frame con cantidad de datos por intervalo de tiempo}
\usage{
info_huecos(xts)
}
\arguments{
\item{xts}{Objeto xts a examinar.}
}
\value{
data.frame con dos columnas: \itemize{
 \item{"dif": }{contiene todos los intervalos de tiempo (en minutos) entre datos, incluyendo el estándar. Si
 solo hay una fila, quiere decir que no hay huecos.}
 \item{"cant": }{cantidad de datos faltantes (huecos) del intervalo de tiempo correspondiente.}
 }
}
\description{
Devuelve data.frame con cantidad de datos del xts segmentado por intervalo de tiempo,
indicando si existen huecos y detallando su cantidad y longitud.
}
\examples{
# ##########DATOS CON HUECOS##########
# #Descarga datos de 1 medidor con datos cada 15 minutos de un mes, sin rellenar huecos (na=NULL)
# datos_con_huecos <- importar_optimum(usuario="usuario_valido", password="password_valido",
#                                      medidores=",CIR0141449301", fecha_inicio = "2019-03-01",
#                                      fecha_fin = "2019-03-31", int_entrada = "15MIN",
#                                      consumo = TRUE, na=NULL)
#
# #Ejecuta info_huecos
# huecos <- info_huecos(datos_con_huecos)
#
# #Viendo el data.frame "huecos", puede verse que los datos descargados tienen 62 huecos de
# #135 minutos. Esto es, dos huecos de 135 minutos por dia.
#
# ##########DATOS SIN HUECOS#########
# #Descarga datos de 1 medidor con datos cada una hora de un mes, sin rellenar huecos (na=NULL)
# datos_sin_huecos <- importar_optimum(usuario="usuario_valido", password="password_valido",
#                                      medidores=",CIR0141437453", fecha_inicio = "2019-03-01",
#                                      fecha_fin = "2019-03-31", int_entrada = "HORA", na=NULL)
#
# #Ejecuta info_huecos
# huecos <- info_huecos(datos_sin_huecos)
#
# #Viendo el data.frame "huecos", puede verse que los datos descargados no tienen huecos, ya que el
# #data.frame solo tiene una fila, correspondiente al intervalo de tiempo estandar entre datos.
}
\author{
Daniel G. Paniagua
}
