% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{pr}
\alias{pr}
\title{Calcula PR a partir de datos de generación real y radiación solar efectiva en el plano del panel}
\usage{
pr(
  Gef,
  int_entrada = "MENSUAL",
  Pn = 1560,
  year = as.POSIXlt(Sys.Date())$year + 1900 - 1
)
}
\arguments{
\item{Gef}{Objeto xts con al menos una columna "Gef" con datos de radiación en el plano del panel FV y
otra columna "gen" con datos de generación de energía de ese sistema. Las unidades de ambas columnas
se deben corresponder.}

\item{int_entrada}{Cadena de caracteres. "MENSUAL" o "DIARIO", según lo sean los datos de entrada.}

\item{Pn}{Numérico. Potencia nominal del sistema en Wp.}

\item{year}{Numérico. Año al que corresponden los datos.}
}
\value{
El objeto xts original con una columna adicional con el PR.
}
\description{
Calcula PR a partir de datos de generación real y radiación solar efectiva en el plano del panel
}
\examples{
# library(OESolar)
#
# #importa datos de generacion cinco medidores de sistema Optimum.
# gen_dia <- importar_optimum(usuario = "usuario_valido", password = "password_valido",
#                             medidores = "GEN1HORA", fecha_inicio = "2019-01-01",
#                             fecha_fin = "2019-12-31", int_entrada = "HORA",
#                             int_salida = "DIARIO", na="INTERPOLAR", mult_us = TRUE)
# nombres <- names(gen_dia)
#
# #Convierte el objeto "list" a una unica tabla donde cada medidor queda en una columna distinta.
# xts.gen_dia <- do.call(merge, gen_dia)
# names(xts.gen_dia) <- nombres
#
# #Agrega una columna con el promedio de los cinco usuarios para cada dia
# xts.gen_dia$gen <- rowMeans(xts.gen_dia)
#
# #Descarga datos de radiacion global en el plano horizontal de Nasa
# rad_dia <- rad_nasa(intervalo = "DIARIO", lonlat = c(-61.60,-32.78),
#                     fechas = c("2019-01-01","2019-12-31"))
# #Obtiene los valores de radiacion en el plano inclinado
# rad_dia_inclin <- rad_inclin(rad_dia, lat = -32.78, beta = 30, int_entrada = "DIARIO")
#
# #Combina la tabla con los datos de generacion con la tabla de datos de radiacion.
# pr <- merge(xts.gen_dia, rad_dia_inclin)
#
# #Aplica la funcion pr, que agrega una columna a la tabla con el valor de PR.
# pr <- pr(pr, int_entrada="DIARIO", Pn=1560)
}
\references{
\itemize{
 \item{""}{van Sark, Wilfried & Reich, Nils & Müller, Björn & Armbruster, Alfons & Kiefer, Klaus & Reise,
 Christian. (2012). Review of PV performance ratio development. 10.13140/2.1.2138.7204. }
}
}
\author{
Daniel G. Paniagua
}
