% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad_inclin.R
\name{rad_inclin}
\alias{rad_inclin}
\title{Calcula el valor de radiación solar en el plano inclinado.}
\usage{
rad_inclin(
  horiz,
  lon = -62.8,
  lat = -32.9,
  beta = 45,
  alfa = 0,
  int_entrada = "MENSUAL",
  iS = 2,
  year = as.POSIXlt(Sys.Date())$year + 1900 - 1
)
}
\arguments{
\item{horiz}{xts con dos columnas: "G0" con radiación en el plano horizontal en kW/m² y
"Ta" con temperatura media en ºC. Para datos mensuales, el índice de fechas en formato yearmon.
Para datos intradiarios, el índice de fechas debe incluir el huso horario correspondiente.}

\item{lon}{Numérico. Longitud de la ubicación del lugar en grados. Solo utilizado para datos
de entrada intradiarios.}

\item{lat}{Numérico. Latitud de la ubicación del lugar en grados. Negativo para hemisferio sur.}

\item{beta}{Numérico. Inclinación del panel en grados.}

\item{alfa}{Numérico. Azimuth en grados. Para hemisferio norte, el sur es cero, es negativo hacia el este
y positivo hacia el oeste. El package solaR no especifica para el hemisferio sur, pero se verificó que el norte es cero.
Tomar mismo criterio que para hemisferio norte para este y oeste.}

\item{int_entrada}{Cadena de caracteres: "MENSUAL", "DIARIO" o "INTRADIARIO" según corresponda a los
datos de radiación y temperatura provistos en horiz.}

\item{iS}{Numérico. Grado de suciedad. Su valor  puede ser 0,1,2,3,4. iS=0 no considera el efecto
de suciedad y reflexión. iS=1 a iS=4 tiene en cuenta el efecto de suciedad y reflexión, donde
iS=1 corresponde a una superficie limpia e iS=4 corresponde a una superficie sucia. El valor por defecto es 2.}

\item{year}{Numérico. Año al que corresponden los datos. Solo utilizado para datos mensuales.
Valor por defecto: año anterior al actual.}
}
\value{
Devuelve el xts \code{horiz} con una columna adicional con la radiación en el plano del panel.
}
\description{
Calcula el valor de radiación solar en el plano inclinado
a partir de la radiación solar en el plano horizontal y la temperatura.
Funciona para datos diarios y datos mensuales.
}
\examples{
# #Calcula valores de radiacion en el plano inclinado para datos diarios de
# #de radiacion solar de Rosario del anio 2019.
#
# rad <- rad_nasa(intervalo = "DIARIO", fechas=c("2019-01-01","2019-12-31"))
# rad_incl <- rad_inclin(rad, beta=28, alfa=11, int_entrada = "DIARIO")
#
# #Calcula valores de radiacion en el plano inclinado para datos mensuales
# #(promediados de los ultimos 30 anios) de radiacion de la ciudad de Rosario, Argentina.
#
# rad <- rad_nasa(intervalo = "MENSUAL", fechas = 2018)
# rad_incl <- rad_inclin(rad, beta=28, alfa=11)
}
\references{
\itemize{
 \item{}{Perpiñán, O, Energía Solar Fotovoltaica, 2015. (http://oscarperpinan.github.io/esf/)}
 \item{}{Perpiñán O (2012). “solaR: Solar Radiation and Photovoltaic Systems with R.” Journal
 of Statistical Software, 50(9), 1–32. http://www.jstatsoft.org/v50/i09/. }
 }
}
\author{
Daniel G. Paniagua
}
