% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad_nasa.R
\name{rad_nasa}
\alias{rad_nasa}
\title{Obtiene datos de radiación de Nasa Power}
\usage{
rad_nasa(
  intervalo = "HISTORICO",
  lonlat = c(-60.8, -32.9),
  fechas = as.POSIXlt(Sys.Date())$year + 1900 - 1
)
}
\arguments{
\item{intervalo}{Una cadena de caracteres.\cr
\code{intervalo="DIARIO"} para datos diarios.\cr
\code{intervalo="MENSUAL"} para datos mensuales de un año especificado en "fechas".\cr
\code{intervalo="HISTORICO"} para promedio mensual sobre período de 30 años.\cr}

\item{lonlat}{Vector con longitud y latitud de la localidad.}

\item{fechas}{Para:
\code{intervalo="DIARIO"}, vector con fecha inicio y fin. Ej: c("2018-01-01", "2018-12-31").\cr
\code{intervalo="MENSUAL"}, especificar un año con número.\cr
\code{intervalo="HISTORICO"}, especificar un año con número. Los datos no corresponderán al año
especificado, sino que será un promedio sobre los últimos 30 años. Este valor se utilizará para asignar
un año a los datos, que será útil para futuras comparaciones con datos de generación y cálculo de PR.
Valor por defecto: año anterior al año en curso.}
}
\value{
Un xts con dos columnas: G0 y Ta (radiación global en el plano horizontal
y temperatura ambiente media). Para "HISTORICO" y "MENSUAL" el formato del índice de fechas es yearmon.
}
\description{
Obtiene datos de radiación solar global en el plano horizontal y
de temperatura de la base de datos Nasa Power.
}
\examples{
# #Devuelve dataframe con datos mensuales (promediados de los ultimos
# #30 anios) de radiacion y temperatura de la ciudad de Rosario, Argentina.
#
# rad <- rad_nasa()
#
# #Devuelve dataframe con datos mensuales de radiacion y temperatura
# #de Rosario, Argentina del 2018.
#
# rad <- rad_nasa(intervalo = "MENSUAL", fechas = 2018)
#
# #Devuelve dataframe con datos diarios de radiacion y temperatura
# #de Marcos Juarez, Argentina del 2019.
#
# rad <- rad_nasa(intervalo = "DIARIO",
#                 lonlat = c(-62.15,-32.7),
#                 fechas = c("2019-01-01","2019-12-31"))
}
\references{
\itemize{
 \item{}{Sparks A (2019). nasapower: NASA-POWER Data from R. R package version 1.1.3,
 https://CRAN.R-project.org/package=nasapower. }
 \item{}{Sparks AH (2018). “nasapower: A NASA POWER Global Meteorology, Surface Solar Energy and Climatology
 Data Client for R.” The Journal of Open Source Software, 3(30), 1035. doi: 10.21105/joss.01035. }
 }
}
\author{
Daniel G. Paniagua
}
