% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rellena_huecos.R
\name{rellena_huecos}
\alias{rellena_huecos}
\title{Rellena huecos de datos faltantes en xts con datos de energía}
\usage{
rellena_huecos(
  xts,
  int_entrada = "HORA",
  na = "INTERPOLAR",
  consumo = FALSE,
  indicar_na = FALSE
)
}
\arguments{
\item{xts}{Objeto xts con datos de energía a rellenar. La columna con los datos debe llamarse "energia".}

\item{int_entrada}{Cadena de caracteres. "15MIN" O "HORA". Corresponde al
intervalo de tiempo entre los datos de entrada. No utilizar datos que mezclen estos valores.}

\item{na}{Cadena de caracteres. "CERO" o "INTERPOLAR", según cómo se quieran rellenar
los huecos: \itemize{
 \item{"CERO": }{rellena todos los huecos con ceros.}
 \item{"INTERPOLAR": }{ rellena huecos interpolando los datos faltantes con los datos anterior y posterior al hueco.
 Si \code{consumo=TRUE}, interpola para todos los horarios. Si \code{consumo=FALSE}, interpola en horarios diurnos
 y completa con cero en horarios nocturnos.}
 }}

\item{consumo}{Lógico. TRUE si los datos son de consumo. Si es FALSE (cuando los datos son de generación),
la interpolación para NA en horario nocturno no se hace y se completa con ceros.}

\item{indicar_na}{Lógico. Si es TRUE, los datos agregados se identifican con un "1" en una columna adicional
llamada "rellenado".}
}
\value{
xts estrictamente regular según el intervalo de tiempo ingresado en \code{int_entrada}.
}
\description{
Rellena huecos de datos faltantes en xts con datos de energía cada 15 minutos o cada 1 hora.
}
\examples{
# ##########DATOS CON HUECOS##########
# #Descarga datos de 1 medidor con datos cada 15 minutos de un mes, sin rellenar huecos (na=NULL)
# datos_con_huecos <- importar_optimum(usuario="usuario_valido",
#                                      password="password_valido",
#                                      medidores=",CIR0141449301", fecha_inicio = "2019-03-01",
#                                      fecha_fin = "2019-03-31", int_entrada = "15MIN",
#                                      consumo = TRUE, na=NULL)
#
# datos_rellenados <- rellena_huecos(xts=datos_con_huecos, int_entrada="15MIN", na="INTERPOLAR",
#                                    consumo=TRUE, indicar_na=TRUE)
#
# info_huecos(datos_con_huecos)
# #62 huecos de 135 minutos (8 datos faltantes de 15 minutos por hueco)
#
# length(datos_rellenados$energia)-length(datos_con_huecos$energia)
# #[1] 496
# #Puede verse que se rellenaron 496 datos (8x62)
}
\author{
Daniel G. Paniagua
}
